tegra fb: Cleanup partially removed fb disable code.
An ioctl to disable fb used to exist and called from
the nvrm daemon code. This was partially removed sometime
back and there are still traces of that code causing some
confusion which is cleaned up by this commit.

ioctl define is left as-is so maintain compatibility
to user mode code.

Change-Id: I3f792d6be075a8754ea448d588e484afc67bd8ba
Reviewed-on: http://git-master/r/878
Reviewed-by: Venkata (Muni) Anda <vanda@nvidia.com>
Tested-by: Venkata (Muni) Anda <vanda@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>