UPSTREAM: arm: msm: smd: fix SMD modem processor sync condition
When booting up we need to wait for the modem processor to
partially boot. This is because the modem processor does
resource allocation for us. If we don't wait the modem won't
honor our requests and we end up crashing or in an unknown
state. This change just formalizes the waiting process.

Signed-off-by: Daniel Walker <c_dwalke@quicinc.com>
(cherry picked from commit 0aec66d493c1476daa88ec56b1076e51f1ceff0b)

Signed-off-by: Benson Leung <bleung@chromium.org>