tegra-nvec: Report keystroke for the key that is used to wake up.
With this change, display comes up with a single key press; but the catch is
that it only works with "Windows" and "F3" keys. This limitation is posed by
Android and not EC.

NOTE that for other than "Windows" and "F3" key presses, the system does wake up,
but an additional "Windows"/"F3" key press is still required to get the display up.

Also, explicitly letting EC know that it should allow any keyboard/mouse event
for wake-up. This is the default behavior, but still wanted to explicitly
specify.

Change-Id:      I2a1e7d90fbbfd810f2334ee3832e193bfb5e45b1
Reviewed-on: http://git-master/r/986
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Tested-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: Mayuresh Kulkarni <mkulkarni@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>