[merge] cleanup config as well as build errors
There are some new options that appear in the configs so i renormalized
the configs to remove the prompting. also i wrapped the use of

tegra_irq_suspend
tegra_irq_resume
tegra_gpio_suspend
tegra_gpio_resume

in CONFIG_PM as the actual function defines are and these references
should match so that they properly follow the configuration file
rules.