Add SMBUS support to the light sensor driver.
This is so we can support it on x86 SMBUS adapters.

This was tested on a board modified to have the tsl2560 on the right bus.
The "lux" file is exported to sysfs by the driver and appears to accurately
reflect changes to ambient light.  That is, covering the sensor with a hand
gives a low value, and the value gradually increases as the hand is moved
away.

Review URL: http://codereview.chromium.org/1955001