tegra: ALSA: Do not queue buffers to the mixer in stopped state
In SNDRV_PCM_TRIGGER_STOP state of the driver, if a buffer gets queued
then it might not be returned back as it never gets rendered.  In this
scenario play_thread never exits as it keep on waiting for receiving
all the buffers back from mixer before exiting.

This is to solve bug where ALSA driver may hang in case browser window
is closed duiring a streaming playback.

In case of START->STOP->START scenario (happens in case of underruns),
without closing the stream, there will not be any buffer loss as buffers
are queued by application at appl_ptr and they remain there, once next
START is called, they will be picked up for the playback.

Signed-off-by: Sandeep Shinde <sashinde@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/2002010