tegra: adding ar6002 wifi driver to chrome kernel
this driver works for the tegra2 and vogue systems for their
wifi controllers. It is built into the kernel. There are seperate
commits to add the firmware required in via and ebuild from the
tegra2-dev-board and vogue public overlays.

There are 2 seperate issues tracking the adding of the required
firmware to the tegra2 and voguet20 overlays, they are issues

2004001 and 1980003 respectively.
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/1996003