gpio: Configuring pinmux and power rail with gpio_request.
When  gpio client uses the direct linux api which is provided by
gpio lib of linux kernel driver, the pinmux and io power rail was
not getting configured.
Configuring the pinux and power rail so that client can use direct
gpio api of linux and need not to use the rm gpio wrapper api.

Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/2104007