tegra harmony: remove the WAR added to turn on disaply after resume
When this code was written, we were not aware of how to turn on harmony's
display using a single key press. Also, after resume if none of the wake-locks
is acquired then system immediately goes to suspend.

To avoid this, we added this WAR to acquire a timed wake lock of 10 sec.
This gives the user a chance to either press a
key/insert-remove usb cable/insert-remove SD card after
which the display turns on.

Now that we have the way to turn on display by any press, this WAR is not
needed.

Change-Id: I26492f01ed8cf07c5cf448c0e1f67dc99251aaf4
Reviewed-on: http://git-master/r/1236
Reviewed-by: Mayuresh Kulkarni <mkulkarni@nvidia.com>
Tested-by: Mayuresh Kulkarni <mkulkarni@nvidia.com>
Reviewed-by: Narendra Damahe <ndamahe@nvidia.com>
Tested-by: Narendra Damahe <ndamahe@nvidia.com>
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/2071026