tegra: Making the GMC pin group to pull up on harmony.
The GMC group is used by UARTD in the harmony and by power on reset, this
group's pullupdown is set to normal. So during kernel boot it throws some
junk character in the serial terminal.

Setting this group pullupdown to pull up resolve this issue.

Change-Id: I28868cd87c88d542c5b29016b7c27ae29cfbc189
Reviewed-on: http://git-master/r/1159
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Gary King <gking@nvidia.com>
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/2130018