fsl udc: Fix cable connect/disconnect checking
On VBUS interrupt irq acknowledges the VBUS interrupt and schedules the work
thread for powering/up down the USB phy and clocks based on the VBUS status.
But, in the work thread instead of checking the VBUS status directly,
rechecking the interrupt status which is already disabled. This is fixed by
checking the vbus status only in the work thread for controlling the power
up/down sequence. Due to the disconnect problem power numbers increased when
there is no USB cable connected to the system.

Change-Id: Id1c5b3cd1f2f195c4dfadf397e1f8650bf9da12b
Reviewed-on: http://git-master/r/1300
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Gary King <gking@nvidia.com>
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/2141001