tegra power: Fix gart suspend/resume operations
Apply proper barriers to suspend/resume operations.
The dmb() in suspend is required to prevent a race
between the next ENTRY_ADDR write and the previous
ENTRY_DATA read. The dsb+outer_sync() in resume()
is needed to ensure that the writes and reads occur
as expected.

Change-Id: I2e66ecce6ceb4f4647bde82cb706b739affeb858
Reviewed-on: http://git-master/r/1291
Reviewed-by: Gary King <gking@nvidia.com>
Reviewed-by: Mayuresh Kulkarni <mkulkarni@nvidia.com>
Tested-by: Mayuresh Kulkarni <mkulkarni@nvidia.com>
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/2143001