CHROMIUM: Hide cursor on first VT.
There are several commits in 2.6.33 that add a
vt.global_cursor_default kernel option that can
be used to hide the cursor on all VTs:

  d9b263528e01bfbaf716b51f38606b3dfe5ac1e9
  f6c06b6807ff9281295989ebad72523865325a4f
  b434a680a29424856e0f40199daa9f65963c7cb4

I think that we want to limit this to VT1,
though, so I'm going with a one-liner change
for now.  When we pull in the commits, we'll
need to decide which route to take (maybe
disable the cursor globally for release builds
but leave it on otherwise).

Signed-off-by: Daniel Erat <derat@chromium.org>

BUG=chromium-os:3224
TEST=tried it

Review URL: http://codereview.chromium.org/2834036