Fix bug with auth retries if AP sends temporary deauth.
If we get sent a DEAUTH from an AP we are actively trying to authenticate to, we shoot ourselves in the foot if we remove state about the base station.  The retry mechanism (rightly) continues and tries to authenticate 3 times regardless of the incoming DEAUTH but there's no way for future AUTH responses to work because the state needed to complete the auth was lost in the STA when the DEAUTH was processed.  An AP is allowed to return DEAUTH as a temporary failure response.  This is allowed by the 802.11 spec under certain situations (one of which is the situation under which chromium-os:3830 was encountered).  This fix drops the DEAUTH packet in the specific situation when this packet is received in the middle of an authentication.

BUG=chromium-os:3830
TEST=Repeated auth/deauth, module load/unload

Review URL: http://codereview.chromium.org/2829024