CHROMIUM: USB: Don't send URBs to devices that can't take them.
This fixes the bug where Qualcomm Gobis use 100% of CPU when autosuspended by
the USB stack. This patch has been sent upstream, but we have no response yet.

TEST=Adhoc
On a laptop with a Gobi, enable autosuspend for the device, then unplug power.
Wait a few seconds; note ksoftirqd using 100% of CPU. Note unpleasant effects on
battery life.

BUG=None
This is related to the (now-closed) 2888010.

Signed-Off-By: Jason Glasgow <jglasgow@chromium.org>
Signed-Off-By: Elly Jones <ellyjones@chromium.org>

Review URL: http://codereview.chromium.org/2861075