CHROMIUM: arm: config: switch arm-generic to versatile
At some point the default config got switched over to the qualcomm
platform. Instead, change it to versatile (which seems to be the qemu
platform of choice for most people).

Note that qemu-system-arm booting is not yet supported, but this is
a required first step to get there.

Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/3077010