CHROMIUM: arm: Remove QSD8650 and Tegra2 configs
The support for qualcomm and nvidia both live in separate branches, no need
to keep the configs in the master branch.

The config changes of other things are because with these gone, there are new
commonalities between existing configs (i.e. it was just an oldconfig run)

Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/3083003