Don't synth deauth msg when doing local_state_change deauth
Currently, the kernel calls a functinon to cleanup internal state
when wpa-supplicant decides it wants to deauth from an access
point when it is "cleaning up" old state.  Unfortunately, this
function also calls wpa_supplicant back with a synthesized
"deauth" packet, because this is the same function that handles
real incoming DEAUTH from a real AP.

There are two possible solutions to this problem.  First,
wpa_supplicant is being stupid by changing its state due to
a DEAUTH from an AP other than the one it wants to be connected
to.  However I think it is equally bogus that the kernel is
synthesizing a DEAUTH from an AP in response to an SME client
that has explicitly stated that it just wants to clean up local
state and not send (or presumably receive) pacakges related to
this operation.

This patch fixes the latter of these two issues.  I presume this
won't make its way upstream in this form, but it does alleviate
the problem.  I'm going to also dig into wpa_supplicant and fix
the problem there, since I consider it broken too.

BUG=chromium-os:5036
TEST=Observed proper behavior via kernel message

Review URL: http://codereview.chromium.org/3050025