CHROMIUM: Add the USB-over-IP driver from incentivespro.com
Quoting from README.CHROMIUM:

This is a copy of the GPL portions of the USB device server from
IncentivesPro. We use it under Chromium OS to facilitate testing of
devices with test harnesses that must run under other operating
systems.

We have not negotiated rights to distribute the userspace; it is
free-as-in-beer without source.

Fetched from http://incentivespro.com/tusbd2.6.31.zip. The rest of
the path is to fit into the directory structure of
http://incentivespro.com/usb-server.tar.gz (which contains the
userland and an earlier version of tusbd)

BUG=5389
TEST=emerge-arm-generic kernel, emerge-x86-generic kernel with and without USE=tusbd_usb_over_ip

This is a slight modification of issue 3139017 removing the config changes

Review URL: http://codereview.chromium.org/3195003