1/8 CHROMIUM: x86: set CONFIG_DEBUG_BUGVERBOSE=y
Although it adds 33kB to our i386 kernel, most developers and distros find
BUGVERBOSE well worth turning on (and it is already on in our arm configs):
it gives the exact FILE and LINE number for any BUG, and a stack backtrace
to panics.  Yes, it was a waste of space while we could not see the output,
but now that ChromeOS is about to make it readable from RAM preserved across
reboot, it becomes useful: CONFIG_DEBUG_BUGVERBOSE=y common to all our configs.

BUG=chromium-os:837
TEST=manual

Signed-off-by: Hugh Dickins <hughd@chromium.org>

Review URL: http://codereview.chromium.org/3087007