9/8 CHROMIUM: ChromeOS x86: set CONFIG_PRESERVED_RAM=y
Default CONFIG_PRESERVED_RAM=y on ChromeOS i386 systems, but n on Chromium
(unlikely to have the appropriate boot firmware), and n on arm (for now),
and n on x86_64 (so we don't need to fork that config yet -
piman's preference).  Later, when we've added a test for our firmware,
we should enable on Chromium, and on x86_64; and later on arm.

(chromeos/config changes separated out from 4/8, as suggested in review:
ideally this should follow 4/8, but I don't know how to insert an extra
--track branch into the middle of the existing --track linked series.

Late mods by piman to update to latest chromeos configs.)

BUG=chromium-os:2269,chromium-os:4721,chromium-os:4722,chromium-os:4723
TEST=build

Signed-off-by: Hugh Dickins <hughd@chromium.org>

Review URL: http://codereview.chromium.org/3157007