CHROMEOS: ath9k: fix beacon loss after bgscan
When we return to the home channel we were never reseting our beacon
timers, this was casued by the fact that the scanning flag was still
on even after we returned to our home channel. There are also other
reasons why we would get a reset and if we are not off channel
we always need to resynch our beacon timers, because a reset will
clear them.

Cc: stable@kernel.org [2.6.32+]
Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>

[Note: To be replaced by cherry-pick.  Fix is not complete]

BUG=chromimum-os:5715
TEST=Manual

Review URL: http://codereview.chromium.org/3348011