CHROMIUM: dm-verity: add chromeos_set_need_recovery on panic
Enables triggering of recovery mode when dm-verity has need to panic.
When we enable a default-panic on error, this will take the user to
recovery mode.  As discussed, we'd like to provide a more user-friendly
option, but this is a good start toward getting back to a usable system.

 Also fix an off-by-one in the per-target error_behavior. (Next up autotests :)

Signed-off-by: Will Drewry <wad@chromium.org>

BUG=chromium-os:5082
TEST=built on x86-generic, imaged with --verity_error_behavior=2 and corrupted the root.  repeated on both cros and non-cros rw and tested with a per-target error.

Change-Id: I88aeeaf56f22e21dee32ed6e1a64ad58a0f02d10

Review URL: http://codereview.chromium.org/3435007