fs/exec.c: provide the correct process pid to the pipe helper
[This patch was just pulled into the -mm tree]

format_corename uses task_tgid_vnr to provide the numeric pid of a
core-dumping process.  For file-based coredumps, this is perfectly
satisfactory.  However, when the core_pattern contains a pipe, the
substituted PID is invalid in the namespace of the core_pattern pipe
helper, the init namespace.

By changing this, any core collector may now find the process in the
init namespace /proc.  This helps with VFS namespacing too since the
mount root is available via /proc.  Unfortunately, it does not help in
cases of more complex namespacing, like net namespaces.  For that, the
helper thread will need to be migrated to the core-dump namespaces.

I have a separate patch series which implements migrating the
____call_usermodehelper thread to the coredump namespace, but it adds a
fair amount of complexity which might be better handled by someone who
understands that code.  I'm happy to mail it out as well though (it
works, but I don't assign a namespaced pid which may open up issues on
its own).

Signed-off-by: Will Drewry <wad@chromium.org>

BUG=chromium-os:6615
TEST=built linus tot; confirmed works with a custom core handler program.  Currently building on x86-generic to test in my vfs/pid namespace utopia.

Review URL: http://codereview.chromium.org/3454004

Change-Id: I953920848581af515fc2d48995ab3c65c1372117