Revert "MOBLIN: fastboot: retry mounting the root fs if we can't find init"
This reverts commit cebaa9ce060546f5d5e13a1ef3158ee858abfbf1.

Turns out that the code is incorrect and will _always_ crash. Luckily we
only execute it when rootfs is already damaged or unavailable, but it's
confusing to keep it in.

The crashes are due to prepare_namespace() being __init, and it is
being called after init code and data has been freed.

Change-Id: Id31676225c58104378e6f23e8b3ce8c729d807f1

BUG=
TEST=

Review URL: http://codereview.chromium.org/3429014