Reduce reacquire timers so that full scan happens sooner after resume from suspend
This modification reduces the re-acquisition time for access points when the
system goes to sleep in one environment and wakes up in a different one.  The
default Linux kernel parameters spend way too long retrying the current AP.
This change modestly tunes down these parameters.  I did tests to make sure
that in a normal suspend case we re-find our AP, and did empirical tests to
confirm that we gain 2 seconds on average with this change.

BUG=none
TEST=Test parameters:

  - Running in WiFi testbed using prototype hardware
  - ./bin/cros_run_wifi_tests.sh --cell=5 -a "test_pat=002*" Roaming
  - This runs network_WiFiRoaming/network_WiFiRoaming.002Suspend
  - The "wait_service_suspend_end_ready" result gives a fairly accurate
    figure of how long it took for the DUT to reacquire the access point
    after returning from suspend and finding out that it has switched
    channels.

With TOT build, the results are, after 5 runs:
   15.943 15.616 15.616 15.847 15.89

With modification of kernel to reduce number of probes and reduce
time waiting for probe responses:
   13.924 13.416 13.51 13.909 13.915

Review URL: http://codereview.chromium.org/3553003