CHROMIUM: config: ext3: default to ordered mode
This should address some of the file corruption issues we have been
seeing. ordered mode is the default on most mainstream linux distros
and is the default for the lucid kernel on which we are based:

http://kernel.ubuntu.com/git?p=ubuntu/ubuntu-lucid.git;a=blob;f=debian.master/config/config.common.ubuntu

ordered mode was recently made the default upstream also:

http://lkml.org/lkml/2010/7/21/176
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=aa32a796389bedbcf1c7714385b18714a0743810

The log message from the upstream commit:

data=writeback mode is dangerous as it leads to higher data loss and stale data
exposure when systems crash. It should not be the default, especially when all
major distros ensure their ext3 filesystems default to ordered mode. Change the
default mode to the safer data=ordered mode, because we should be caring far
more about avoiding stale data exposure than performance.

BUG=none
TEST=Verified no impact on boot performance.
Machine is has been up for 21 hours without issue.

Change-Id: I49f80fc736dc1435404d34a31fe6cb5136c7973e

Review URL: http://codereview.chromium.org/3557011