CHROMIUM: seccomp: allow TSC access
We use the seccomp sandbox and side channel attacks via the tsc are
not a specific concern (caused by tsc access).  This will allow fast
rdtsc access from inside the jails and if there is a specific concern,
it can be disabled prior to entering the seccomp jail via prctl.

This will result in an immediate speed increase for every rdtsc call
in the seccomp sandbox because it relies on catching SIGSEGV to emulate
a working RDTSC call:
http://codesearch.google.com/codesearch/p?hl=en#OAMlx_jo-ck/src/seccompsandbox/sandbox.cc&exact_package=chromium&l=126

TEST=builds x86-generic and rdtsc works
BUG=chromium-os:6502

Change-Id: Ib44483cb7f342c2aa26901f878b20f9f2575a420

Review URL: http://codereview.chromium.org/3672005